+++ src/main/java/cl/desagen/chilquinta/repositories/NormaRepository.java
@@ -3,6 +3,6 @@ package cl.desagen.chilquinta.repositories;
 import cl.desagen.chilquinta.entities.NormaEntity;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface NormaRepository extends PagingAndSortingRepository<NormaEntity, Long> {
+public interface NormaRepository extends PagingAndSortingRepository<NormaEntity, Integer> {
 
 }
